Move GeoPointActivity to geo module
