+++ androidshared/src/main/res/color/color_on_surface_medium_emphasis.xml
