+++ collect_app/src/main/java/org/odk/collect/android/widgets/utilities/ActivityGeoDataRequester.java
@@ -6,7 +6,7 @@ import android.content.Intent;
 import android.os.Bundle;
 
 import org.javarosa.form.api.FormEntryPrompt;
-import org.odk.collect.android.activities.GeoPointActivity;
+import org.odk.collect.geo.GeoPointActivity;
 import org.odk.collect.android.activities.GeoPointMapActivity;
 import org.odk.collect.android.activities.GeoPolyActivity;
 import org.odk.collect.android.listeners.PermissionListener;
