+++ geo/src/main/res/layout/geopoint_dialog.xml
